+++ sdk/communication/azure-communication-sms/src/samples/java/com/azure/communication/sms/samples/quickstart/ReadmeSamples.java
@@ -2,13 +2,10 @@
 // Licensed under the MIT License.
 package com.azure.communication.sms.samples.quickstart;
 
-import java.security.InvalidKeyException;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 
 import com.azure.communication.common.PhoneNumber;
-import com.azure.communication.common.CommunicationClientCredential;
 import com.azure.communication.sms.SmsClient;
 import com.azure.communication.sms.SmsClientBuilder;
 import com.azure.communication.sms.models.SendSmsOptions;
@@ -29,22 +26,13 @@ public class ReadmeSamples {
 
         // Instantiate the http client
         HttpClient httpClient = new NettyAsyncHttpClientBuilder().build();
-        
-        CommunicationClientCredential credential = null;
-        try {
-            credential = new CommunicationClientCredential(accessKey);
-        } catch (NoSuchAlgorithmException e) {
-            System.out.println(e.getMessage());
-        } catch (InvalidKeyException e) {
-            System.out.println(e.getMessage());
-        }
-        
+
         // Create a new SmsClientBuilder to instantiate an SmsClient
         SmsClientBuilder smsClientBuilder = new SmsClientBuilder();
 
         // Set the endpoint, access key, and the HttpClient
         smsClientBuilder.endpoint(endpoint)
-            .credential(credential)
+            .accessKey(accessKey)
             .httpClient(httpClient);
 
         // Build a new SmsClient
@@ -67,5 +55,18 @@ public class ReadmeSamples {
             options /* Optional */);
 
         System.out.println("MessageId: " + response.getMessageId());
+    }
+
+    public void createCommunicationIdentityClientWithConnectionString() {
+        // Create an HttpClient builder of your choice and customize it
+        HttpClient httpClient = new NettyAsyncHttpClientBuilder().build();
+
+        // Your can find your connection string from your resource in the Azure Portal
+        String connectionString = "<connection_string>";
+
+        SmsClient smsClient = new SmsClientBuilder()
+            .connectionString(connectionString)
+            .httpClient(httpClient)
+            .buildClient();
     }
 }