+++ sdk/servicebus/azure-messaging-servicebus/src/test/java/com/azure/messaging/servicebus/ServiceBusReceiverAsyncClientTest.java
@@ -15,7 +15,9 @@ import com.azure.core.credential.TokenCredential;
 import com.azure.core.util.ClientOptions;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.messaging.servicebus.ServiceBusClientBuilder.ServiceBusReceiverClientBuilder;
-import com.azure.messaging.servicebus.administration.models.DeadLetterOptions;
+import com.azure.messaging.servicebus.models.AbandonOptions;
+import com.azure.messaging.servicebus.models.CompleteOptions;
+import com.azure.messaging.servicebus.models.DeadLetterOptions;
 import com.azure.messaging.servicebus.implementation.DispositionStatus;
 import com.azure.messaging.servicebus.implementation.MessageWithLockToken;
 import com.azure.messaging.servicebus.implementation.MessagingEntityType;
@@ -23,9 +25,11 @@ import com.azure.messaging.servicebus.implementation.ServiceBusAmqpConnection;
 import com.azure.messaging.servicebus.implementation.ServiceBusConnectionProcessor;
 import com.azure.messaging.servicebus.implementation.ServiceBusManagementNode;
 import com.azure.messaging.servicebus.implementation.ServiceBusReactorReceiver;
+import com.azure.messaging.servicebus.models.DeferOptions;
 import com.azure.messaging.servicebus.models.ReceiveMode;
 import org.apache.qpid.proton.amqp.messaging.Rejected;
 import org.apache.qpid.proton.amqp.transport.DeliveryState.DeliveryStateType;
+import org.apache.qpid.proton.engine.SslDomain;
 import org.apache.qpid.proton.message.Message;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -145,7 +149,8 @@ class ServiceBusReceiverAsyncClientTest {
 
         ConnectionOptions connectionOptions = new ConnectionOptions(NAMESPACE, tokenCredential,
             CbsAuthorizationType.SHARED_ACCESS_SIGNATURE, AmqpTransportType.AMQP, new AmqpRetryOptions(),
-            ProxyOptions.SYSTEM_DEFAULTS, Schedulers.boundedElastic(), CLIENT_OPTIONS);
+            ProxyOptions.SYSTEM_DEFAULTS, Schedulers.boundedElastic(), CLIENT_OPTIONS,
+            SslDomain.VerifyMode.VERIFY_PEER_NAME);
 
         when(connection.getEndpointStates()).thenReturn(endpointProcessor);
         endpointSink.next(AmqpEndpointState.ACTIVE);
@@ -270,46 +275,6 @@ class ServiceBusReceiverAsyncClientTest {
     }
 
     /**
-     * Verifies that we error if we try to settle a message with null transaction.
-     */
-    @ParameterizedTest
-    @EnumSource(DispositionStatus.class)
-    void settleWithNullTransaction(DispositionStatus dispositionStatus) {
-        // Arrange
-        ServiceBusTransactionContext nullTransaction = null;
-        when(connection.getManagementNode(ENTITY_PATH, ENTITY_TYPE)).thenReturn(Mono.just(managementNode));
-        when(managementNode.updateDisposition(any(), eq(dispositionStatus), isNull(), isNull(), isNull(),
-            isNull(), isNull(), isNull()))
-            .thenReturn(Mono.delay(Duration.ofMillis(250)).then());
-        when(receivedMessage.getLockToken()).thenReturn("mylockToken");
-
-        final Mono<Void> operation;
-        switch (dispositionStatus) {
-            case DEFERRED:
-                operation = receiver.defer(receivedMessage, null, nullTransaction);
-                break;
-            case ABANDONED:
-                operation = receiver.abandon(receivedMessage, null, nullTransaction);
-                break;
-            case COMPLETED:
-                operation = receiver.complete(receivedMessage, nullTransaction);
-                break;
-            case SUSPENDED:
-                operation = receiver.deadLetter(receivedMessage, new DeadLetterOptions(), nullTransaction);
-                break;
-            default:
-                throw new IllegalArgumentException("Unrecognized operation: " + dispositionStatus);
-        }
-
-        StepVerifier.create(operation)
-            .expectError(NullPointerException.class)
-            .verify();
-
-        verify(managementNode, never()).updateDisposition(any(), eq(dispositionStatus), isNull(), isNull(),
-            isNull(), isNull(), isNull(), isNull());
-    }
-
-    /**
      * Verifies that we error if we try to settle a message with null transaction-id.
      */
     @ParameterizedTest
@@ -317,6 +282,7 @@ class ServiceBusReceiverAsyncClientTest {
     void settleWithNullTransactionId(DispositionStatus dispositionStatus) {
         // Arrange
         ServiceBusTransactionContext nullTransactionId = new ServiceBusTransactionContext(null);
+
         when(connection.getManagementNode(ENTITY_PATH, ENTITY_TYPE)).thenReturn(Mono.just(managementNode));
         when(managementNode.updateDisposition(any(), eq(dispositionStatus), isNull(), isNull(), isNull(),
             isNull(), isNull(), isNull()))
@@ -326,16 +292,16 @@ class ServiceBusReceiverAsyncClientTest {
         final Mono<Void> operation;
         switch (dispositionStatus) {
             case DEFERRED:
-                operation = receiver.defer(receivedMessage, null, nullTransactionId);
+                operation = receiver.defer(receivedMessage, new DeferOptions().setTransactionContext(nullTransactionId));
                 break;
             case ABANDONED:
-                operation = receiver.abandon(receivedMessage, null, nullTransactionId);
+                operation = receiver.abandon(receivedMessage, new AbandonOptions().setTransactionContext(nullTransactionId));
                 break;
             case COMPLETED:
-                operation = receiver.complete(receivedMessage, nullTransactionId);
+                operation = receiver.complete(receivedMessage, new CompleteOptions().setTransactionContext(nullTransactionId));
                 break;
             case SUSPENDED:
-                operation = receiver.deadLetter(receivedMessage, new DeadLetterOptions(), nullTransactionId);
+                operation = receiver.deadLetter(receivedMessage, new DeadLetterOptions().setTransactionContext(nullTransactionId));
                 break;
             default:
                 throw new IllegalArgumentException("Unrecognized operation: " + dispositionStatus);