+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,12 +1,14 @@
 # Release History
 
-## 12.9.0-beta.1 (Unreleased)
+## 12.9.0-beta.2 (Unreleased)
+- Added support to specify whether or not a pipeline policy should be added per call or per retry.
+
+## 12.9.0-beta.1 (2020-10-01)
 - Added support for the 2020-02-10 service version.
 - Added support to specify Arrow Output Serialization when querying a blob. 
 - Added support to undelete a container. 
 - Added support to set BlobParallelUploadOptions.computeMd5 so the service can perform an md5 verification.
 - Added support to specify block size when using BlobInputStream.
-
 - Added support to specify block size when using BlobInputStream.
 - Fixed a bug where users could not download more than 5000MB of data in one shot in the downloadToFile API.
 - Fixed a bug where the TokenCredential scope would be incorrect for custom URLs.